% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPRestimation.R
\name{caretFolds}
\alias{caretFolds}
\title{Create cross-validation folds (from caret package)}
\usage{
caretFolds(y, k = 10, list = TRUE, returnTrain = FALSE)
}
\arguments{
\item{y}{outcome vector, i.e. the indicator labels for each instance}

\item{k}{number of folds to create}

\item{list}{logical; should folds be returned as a list or matrix?}
}
\description{
This is the caret package's "createFolds" function, copied here
because this is the only function used from the caret package,
so it seemed excessive to list caret as a dependency.
}
