% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPRestimation.R
\name{convertToPercentile}
\alias{convertToPercentile}
\title{Transform classifier scores to percentiles}
\usage{
convertToPercentile(scores)
}
\arguments{
\item{scores}{A vector of classifier scores.}
}
\value{
A vector of percentiles (as decimals rather than percentages), the
same length as \code{scores}.
}
\description{
Given a vector of classifier scores \code{convertToPercentile}
returns the empirical CDF evaluated at each of the scores.
}
