% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{CreateDataForSetting}
\alias{CreateDataForSetting}
\title{Generate synthetic data based on Beta distributions.}
\usage{
CreateDataForSetting(
  n = 60000,
  test_n = 10000,
  min_cases = 150,
  unif_cond = TRUE,
  setting,
  seed
)
}
\arguments{
\item{n}{Number of total objects, including training and testing.}

\item{test_n}{Number of testing objects}

\item{min_cases}{Minimal number of positive cases in each class.}

\item{unif_cond}{A logic value, whether using purely uniform conditional probablity.
A FALSE value accounts for the joint probablity to take into account the \code{min_cases}.}

\item{setting}{The simulation setting. There are 11 hard-coded setting.}

\item{seed}{Random seed.}
}
\value{
A list with four elements.
\describe{
\item{labels}{A K x n matrix containing labels for n objects across K classes.}
\item{scores}{A K x n matrix containing simulated classifier scoresfor n objects across K classes.}
\item{seed}{Random seed.}
\item{node_info}{A K x 6 matrix containing the information of K classes: the first columns is the
the parent of the node, the second column is the conditional probablity \code{Pr(Y_i|Y_{pa(i)})},
the third and the columns are two parameters of the null Beta distribution, the fifth and the
sixth columns are two parameters of the alternative Beta distribution.}
}
}
\description{
Generate synthetic data based on Beta distributions.
}
