% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPRestimation.R
\name{estimateLPR}
\alias{estimateLPR}
\title{Estimate LPR}
\usage{
estimateLPR(
  scores,
  labels,
  newdata = NULL,
  num_bags,
  method = c("kernel", "spline"),
  ...
)
}
\arguments{
\item{scores}{Numeric vector containing classifier scores for each observation}

\item{labels}{A 0-1 vector containing the true status of each observation}

\item{newdata}{An optional numeric vector containing scores for a new data
set, e.g. a test set. If omitted, then LPRs for \code{scores}
are estimated by leave-one-out cross-validation.}

\item{num_bags}{Integer indicating the number of bagging steps. If 0,
no bagging is done.}

\item{method}{Method used to estimate the precision function}

\item{...}{Additional parameters to pass to the function for estimating the
precision (e.g., additional arguments to fitKernelLPR).}
}
\description{
This function estimates the LPR on a data set.
The user has the option of providing two data sets (one to use as training
and the other as a test set). Otherwise, if only one data set is provided,
the function will estimate LPR values for the provided data set using
leave one out CV.
}
\details{
Users have the option of using bagging to produce more robust LPR estimates,
and can choose whether to fit the precision function using a local
polynomial kernel smoother or a spline.
}
