% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPRestimation.R
\name{estimatePrecision}
\alias{estimatePrecision}
\title{Estimate the precision function}
\usage{
estimatePrecision(scores, labels, plot = FALSE)
}
\arguments{
\item{scores}{A vector of classifier scores.}

\item{labels}{True labels for each of the instances in \code{scores}.}

\item{plot}{If TRUE, generates a plot of the precision function.}
}
\value{
A list with three elements.
\describe{
\item{u}{Empirical CDF evaluated at each score.}
\item{prec}{Empirical precision estimates for each score.}
\item{size}{Sample size used to estimate each precision value.}
}
}
\description{
Given a vector of classifier scores and true labels,
\code{estimatePrecision} estimates the precision as a function of the
lower tail probabilities (cdf) of the scores.
}
