% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPRestimation.R
\name{fitKernelLPR}
\alias{fitKernelLPR}
\title{Estimate LPR using local polynomial smoothing}
\usage{
fitKernelLPR(
  train_scores,
  train_labels,
  test_scores = NULL,
  param = NULL,
  weights = NULL,
  nfold = 5
)
}
\arguments{
\item{train_scores}{A vector of classifier scores to use as the training set}

\item{train_labels}{True labels for the instances in the training set}

\item{test_scores}{A vector of classifier scores to obtain LPRs for. If
this is set to NULL, then the function only returns the
best smoothing parameter chosen from an nfold CV on
the training data.}

\item{param}{Smoothing parameter for the spline fit; if NULL, estimate
using \code{nfold} cross validation. If a non-NULL is
provided and test_scores = NULL, then this function
trivially returns the provided value of param.}

\item{nfold}{Use \code{nfold} cross validation to estimate param,
if not given. By default, 5-fold CV is done.}
}
\value{
A list with three elements.
\describe{
\item{u}{Empirical CDF evaluated at each score.}
\item{prec}{Empirical precision estimates for each score.}
\item{size}{Sample size used to estimate each precision value.}
}
}
\description{
Given a vector of classifier scores and true labels,
\code{fitSplineLPR} fits a local quadratic smoother to
the precision function on training data, and provides predictions
on a test set of classifier scores.
}
