% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPRestimation.R
\name{kernelloss}
\alias{kernelloss}
\title{Compute mean squared error loss, calculated per fold (for cv)}
\usage{
kernelloss(param, x, y, folds, w)
}
\arguments{
\item{param}{bandwidth parameter for kernel smoother}

\item{x}{a vector giving the values of the predictor variable}

\item{y}{responses}

\item{folds}{a list of vectors, each giving the indices for each CV fold}
}
\description{
Compute mean squared error loss, calculated per fold (for cv)
}
