% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPRestimation.R
\name{predictNewLPR}
\alias{predictNewLPR}
\title{Estimate LPR on a new data set using bagging}
\usage{
predictNewLPR(
  train_scores,
  train_labels,
  test_scores,
  num_bags,
  method = c("kernel", "spline"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{train_scores}{Numeric vector containing classifier scores for each observation}

\item{train_labels}{A 0-1 vector containing the true status of each observation}

\item{test_scores}{Numeric vector containing the scores for the test set.}

\item{num_bags}{Integer indicating the number of bagging steps. If 0,
no bagging is done.}

\item{method}{Method used to estimate the precision function}

\item{verbose}{If TRUE, prints out at most five progress messages
during LPR estimation using bagging,
since this can be slow when the
number of bags or observations is large.}

\item{...}{Additional parameters to pass to the function for estimating the
precision function (e.g., fitKernelLPR)}
}
\description{
This function estimates the LPR given training data and a new test set.
Users have the option of using bagging to produce more robust LPR estimates,
and can choose whether to fit the precision function using a local
polynomial kernel smoother or a spline.
}
