% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierRank.R
\name{Sorting}
\alias{Sorting}
\title{Sort a vector of values using naive sorting or HierRank.}
\usage{
Sorting(scores, graph_structure, sorting_method = "value")
}
\arguments{
\item{scores}{A K x n matrix}

\item{graph_structure}{A list that indicates the parents of each node.}

\item{sorting_method}{The sorting method, including 'value' (naive sorting in descending order), 'HierRank'.}
}
\value{
A list.
\describe{
\item{ordering}{The resulting ranking.}
\item{values}{The transposed \code{scores}.}
\item{block_ordering}{The ordered blocks in HierRank method (only for 'HierRank').}
}
}
\description{
Sort a vector of values using naive sorting or HierRank.
}
