% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPRestimation.R
\name{splineloss}
\alias{splineloss}
\title{Define mean squared error loss function, calculated per fold (for cv)}
\usage{
splineloss(param, x, y, folds, w = NULL)
}
\arguments{
\item{param}{smoothing parameter, typically in (0, 1]}

\item{x}{a vector giving the values of the predictor variable}

\item{y}{responses}

\item{folds}{a list of vectors, each giving the indices for each CV fold}
}
\description{
Define mean squared error loss function, calculated per fold (for cv)
}
